pyproject.toml
[project]
name = "sdlc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.1",
    "gitpython>=3.1.46",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "loadenv>=0.1.1",
    "pinecone>=8.1.0",
    "google-genai>=1.70.0",
    "openai>=2.30.0",
    "langchain-openai>=1.1.12",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]
